+++ app/Main.hs
@@ -42,6 +42,9 @@ animation = sinceStart >-> proc time -> do
   flag <- switchFlags 0.2 -< time
   returnA -< brightness <> flag
 
+traverseLed :: (Monad m, TimeDomain (Time cl), Diff (Time cl) ~ Double) => ClSF m cl () StatePatch
+traverseLed = sinceStart >-> arr (\t -> (mempty :: StatePatch) { stateBri = Just 255, stateSeg = Just [segment (floor (t / 3.0)) (floor (t / 3.0) + 1 ) [255, 255, 255]]})
+
 main :: IO ()
 main = do
   hSetBuffering stdout NoBuffering
@@ -56,6 +59,7 @@ main = do
       putStrLn ""
       putStrLn "1: Change URL for WLED device."
       putStrLn "2: Run octolamp flags demo."
+      putStrLn "3: Traverse single LED."
       putStrLn "q: Quit."
       putStrLn ""
       putStr $ "(" ++ wledUrl ++ ") > "
@@ -68,21 +72,29 @@ main = do
         "2" -> do
           putStrLn ""
           putStrLn "Press [Enter] to stop demonstration."
-          lampState <- getLampState wledUrl
-          case lampState of
-            Left errMsg -> putStrLn errMsg
-            Right initialState -> do
-              void $ runExceptT $ flow $ waitForEnter @@ StdinClock |@| (animation >-> arrMCl (liftIO . void . setLampState wledUrl)) @@ liftClock @IO @(ExceptT ()) (waitClock @100)
-              Right currentState <- getLampState wledUrl
-              _ <- setLampState wledUrl (diff currentState initialState)
-              -- validate that initial state is restored
-              Right restoredState <- getLampState wledUrl
-              putStrLn $ "Initial state is restored: " <> show (initialState == restoredState)
+          localLampState wledUrl $ void $ runExceptT $ flow $ waitForEnter @@ StdinClock |@| (animation >-> arrMCl (liftIO . void . setLampState wledUrl)) @@ liftClock @IO @(ExceptT ()) (waitClock @100)
+          mainLoop wledUrl
+        "3" -> do
+          putStrLn ""
+          putStrLn "Press [Enter] to stop traversal."
+          localLampState wledUrl $ void $ runExceptT $ flow $ waitForEnter @@ StdinClock |@| (traverseLed >-> arrMCl (liftIO . void . setLampState wledUrl)) @@ liftClock @IO @(ExceptT ()) (waitClock @100)
           mainLoop wledUrl
         "q" -> pure ()
         "Q" -> pure ()
         _ -> do
           putStrLn "Could not recognize option."
           mainLoop wledUrl
+    localLampState :: String -> IO () -> IO ()
+    localLampState wledUrl action = do
+      lampState <- getLampState wledUrl
+      case lampState of
+        Left errMsg -> putStrLn errMsg
+        Right initialState -> do
+          action
+          Right currentState <- getLampState wledUrl
+          _ <- setLampState wledUrl (diff currentState initialState)
+          -- validate that initial state is restored
+          Right restoredState <- getLampState wledUrl
+          putStrLn $ "Initial state is restored: " <> show (initialState == restoredState)
     defaultWledUrl :: String
     defaultWledUrl = "http://192.168.178.34"